refactor: merge rgt example files into utils.py + train.py
Merge rgt_mappings.py into utils.py (class_maps dict), merge
rgt_pretrain.py and rgt_supervised.py into train.py. This follows
the pattern used by other examples (single utils + train pair).
All logic preserved unchanged.
